Remove every file that I possibly can
This commit trims every bit of code that I
possibly can out of the puppet_openstack_buider.

The intention is not that this code should not be here
or that it is not valid. It is being removed b/c I
want to reset this repo, starting with only the most
basic deployment examples, then figuring out how to
add more without creating a lot of weight to the builder
project.

Code removed in this commit may be replaced as deemed necessary.